chi v4, start by cleaning out older Go version patch code
