+++ .travis.yml
@@ -1,7 +1,6 @@
 language: go
 
 go:
-  - 1.9.x
   - 1.10.x
   - 1.11.x
